Prepare for onsets_frames_transcription checkout and pull from fork root
