--- magenta/models/polyamp/onsets_frames_transcription_create_dataset_maestro.py
@@ -1,37 +1,0 @@
-# Copyright 2020 The Magenta Authors.
-# Modifications Copyright 2020 Jack Spencer Smith.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-r"""Beam job for creating transcription dataset from MAESTRO."""
-
-from magenta.models.polyamp import configs
-from magenta.models.polyamp import create_dataset_maestro
-from magenta.models.polyamp import dataset_reader
-from absl import app
-
-
-def main(argv):
-    del argv
-
-    create_dataset_maestro.pipeline(
-        configs.CONFIG_MAP, configs.DATASET_CONFIG_MAP, dataset_reader.preprocess_example,
-        dataset_reader.input_tensors_to_example)
-
-
-def console_entry_point():
-    app.run(main)
-
-
-if __name__ == '__main__':
-    console_entry_point()
